requirements.txt
decord==0.6.0
ffmpeg_python==0.2.0
flash_attn==2.7.4.post1
huggingface_hub==0.30.2
imageio==2.37.0
loguru==0.7.3
numpy==2.2.5
opencv_python==4.11.0.86
Pillow==11.2.1
pymilvus==2.5.6
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
torch==2.6.0
tqdm==4.67.1
transformers==4.51.2
